refactored into components, added Presenter and Presentable
