+++ app/components/kadmin/finder.rb
@@ -1,6 +1,9 @@
+# frozen_string_literal: true
 module Kadmin
   class Finder
-    # @return [Kadmin::Pagination::Pager] the pager to use (if any)
+    include Kadmin::Presentable
+
+    # @return [Kadmin::Pager] the pager to use (if any)
     attr_reader :pager
 
     # @return [Array<Kadmin::Finder::Filter>] array of filters applied to the finder
@@ -43,7 +46,7 @@ module Kadmin
       size = size.to_i
 
       if size.positive? && offset >= 0
-        @pager = Kadmin::Pagination::Pager.new(size: size, offset: offset)
+        @pager = Kadmin::Pager.new(size: size, offset: offset)
       end
 
       return self
