+++ app/views/layouts/modular/application.html.erb
@@ -44,7 +44,7 @@
 						<nav class="menu">
 							<ul class="nav metismenu" id="sidebar-menu">
 								<% Kadmin.config.navbar_items.each do |item| %>
-								<%= item.present(self).html_safe %>
+								<%= item.present(self).render %>
 								<% end %>
 							</ul>
 						</nav>
