--- lib/kadmin/navigation.rb
@@ -1,7 +1,0 @@
-require 'kadmin/navigation/link'
-require 'kadmin/navigation/section'
-
-module Kadmin
-  module Navigation
-  end
-end
