--- lib/kadmin/pagination.rb
@@ -1,7 +1,0 @@
-require 'kadmin/pagination/pager'
-require 'kadmin/pagination/presenter'
-
-module Kadmin
-  module Pagination
-  end
-end
